Push geometry event updates into feature
Rather than waiting until the end of editing. This
doesn't actually make any difference right now; eventually
it might be useful for the form to see a more up-to-date
geometry in case some fields of the form rely on the
geometry.

This isn't as reactive as it could be. The geometry is only
pushed after each geometry editor interaction finishes, but
not "live" as the interaction is happening. I don't know if
we could really achieve that without a lot of effort at the
API level.