Add two-part toolbar with geometry buttons
The sizing logic is not great with this toolbar. If the two
sub-toolbars (the "control one" -- stop, discard, att/geom --
and the geometry one) have different numbers of buttons then
the whole thing will shift around.

I think we will need to do something more complicated to fix that.
Possibly getting an intrinsic measurement of the width of the control
toolbar and using that as a minimum width of the geometry toolbar,
in combination with making it scrollable.