Prevent jarring when switching between subtoolbars
The jarring is caused by the fact that the different subtoolbars
have different widths, which cause the overall toolbar to have a
different width depending on which subtoolbar is being displayed.
That then causes jarring because the content is centered horizontally.

We do still want the content to be centered horizontally but without jarring.
We can kind of achieve this by inserting another row after the horizontal-
centering which takes up a fixed width on the screen (80% in this case).
We make this row not have a horizontal centering, meaning that the toolbar
fills from the left and there is no jarring but it's not completely
pinned to the left because it's contrained to be within the centre 80%
of the screen.

This is ok but it doesn't work great on wide screens because the toolbar
is still very far to the left.