Make both toolbars always the same size
Currently there is still a bit of jarring on the RHS because
the toolbars may have different numbers of buttons. Also the current
solution pushes the toolbar over to the left on big screens.

We can instead remove the extra row added in the previous solution and
pursue a different approach, which is to size the geometry toolbar
based on the size of the control toolbar. This usually isn't possible
in Compose but we can do it with SubComposeLayout.

If the geometry toolbar is bigger than the control toolbar then
it will scroll to allow the user to access the extra buttons. In
my opinion this is not very intuitive. It seems like a fundamental
flaw with the design of the toolbar UI.

Note that this solution could probably be optimised to avoid
unnecessary measurements.