Update dependencies (#684)
* Update dependencies

* fix compilation issue

* update gradle plugin

* fix lint error

Using Modifier.then with a Modifier factory function with an implicit receiver More... (⌘F1)
Inspection info: Calling a Modifier factory function with an implicit receiver inside Modifier.then will result in the receiver (this) being added twice to the chain. For example, fun Modifier.myModifier() = this.then(otherModifier()) - the implementation of factory functions such as Modifier.otherModifier() will internally call this.then(...) to chain the provided modifier with their implementation. When you expand this.then(otherModifier()), it becomes: this.then(this.then(OtherModifierImplementation)) - so you can see that this is included twice in the chain, which results in modifiers such as padding being applied twice, for example. Instead, you should either remove the then() and directly chain the factory function on the receiver, this.otherModifier(), or add the empty Modifier as the receiver for the factory, such as this.then(Modifier.otherModifier())  Issue id: SuspiciousModifierThen Vendor: Jetpack Compose Identifier: androidx.compose.ui Feedback: https://issuetracker.google.com/issues/new?component=612128

* update gradle plugin version

* Update gradle to latest

* update compileSdk and minSdk to use libs.version for microapps-lib