retrieve artifactory info from local.properties (#808)
* Remove artifactory credentials from the gradle.properties

* update settings.gradle for get artifactory info

Add logic that will try to retrieve the artifactory information from the
command line, then from the local.properties file, and fail if it cannot
find the required information.

* update build.gradle for setting artifactory info

Add logic that will try to retrieve the artifactory information from the
command line, then from the local.properties file, and fail if it cannot
find the required information. Then set them as project variables to be
used by other tasks.

* ci: update publish.sh to require artifactory url

Update the publish script to require the artifactory url as an environment
variable

* gunther: suggestions